feat(merlin-www-top-stories): add ad rendering in js
